requirements.txt
path==16.7.1
playsound==1.2.2
progressbar2==4.2.0
pygame==2.3.0
pyserial==3.5
python-utils==3.7.0
serial-device2==2.0.2
typing_extensions==4.7.1
